Allow reading label configuration from a hosted repository file
