+++ forge/src/main/java/org/openjdk/skara/forge/LabelConfiguration.java
@@ -22,130 +22,11 @@
  */
 package org.openjdk.skara.forge;
 
-import org.openjdk.skara.json.JSONValue;
-import org.openjdk.skara.json.JSONObject;
-
 import java.nio.file.Path;
-import java.util.*;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
+import java.util.Set;
 
-public class LabelConfiguration {
-    private final Map<String, List<Pattern>> matchers;
-    private final Map<String, List<String>> groups;
-    private final Set<String> extra;
-    private final Set<String> allowed;
-
-    private LabelConfiguration(Map<String, List<Pattern>> matchers, Map<String, List<String>> groups, Set<String> extra) {
-        this.matchers = Collections.unmodifiableMap(matchers);
-        this.groups = Collections.unmodifiableMap(groups);
-        this.extra = Collections.unmodifiableSet(extra);
-
-        var allowed = new HashSet<String>();
-        allowed.addAll(matchers.keySet());
-        allowed.addAll(groups.keySet());
-        allowed.addAll(extra);
-        this.allowed = Collections.unmodifiableSet(allowed);
-    }
-
-    public static class Builder {
-        private final Map<String, List<Pattern>> matchers = new HashMap<>();
-        private final Map<String, List<String>> groups = new HashMap<>();
-        private final Set<String> extra = new HashSet<>();
-
-        public Builder addMatchers(String label, List<Pattern> matchers) {
-            this.matchers.put(label, matchers);
-            return this;
-        }
-
-        public Builder addGroup(String label, List<String> members) {
-            groups.put(label, members);
-            return this;
-        }
-
-        public Builder addExtra(String label) {
-            extra.add(label);
-            return this;
-        }
-
-        public LabelConfiguration build() {
-            return new LabelConfiguration(matchers, groups, extra);
-        }
-    }
-
-    public static Builder builder() {
-        return new Builder();
-    }
-
-    public static LabelConfiguration fromJSON(JSONValue json) {
-        var builder = builder();
-        if (json.contains("matchers")) {
-            var fields = json.get("matchers").fields();
-            var matchers = fields.stream()
-                                 .collect(Collectors.toMap(JSONObject.Field::name,
-                                                           field -> field.value()
-                                                                         .stream()
-                                                                         .map(JSONValue::asString)
-                                                                         .map(Pattern::compile)
-                                                                         .collect(Collectors.toList())));
-            matchers.forEach(builder::addMatchers);
-        }
-        if (json.contains("groups")) {
-            var fields = json.get("groups").fields();
-            var groups = fields.stream()
-                               .collect(Collectors.toMap(JSONObject.Field::name,
-                                                         field -> field.value()
-                                                                       .stream()
-                                                                       .map(JSONValue::asString)
-                                                                       .collect(Collectors.toList())));
-            groups.forEach(builder::addGroup);
-        }
-        if (json.contains("extra")) {
-            var extra = json.get("extra").stream()
-                                         .map(JSONValue::asString)
-                                         .collect(Collectors.toList());
-            extra.forEach(builder::addExtra);
-        }
-        return builder.build();
-    }
-
-    public Set<String> label(Set<Path> changes) {
-        var labels = new HashSet<String>();
-        for (var file : changes) {
-            for (var label : matchers.entrySet()) {
-                for (var pattern : label.getValue()) {
-                    var matcher = pattern.matcher(file.toString());
-                    if (matcher.find()) {
-                        labels.add(label.getKey());
-                        break;
-                    }
-                }
-            }
-        }
-
-        var ret = new HashSet<>(labels);
-        // If the current labels matches at least two members of a group, use the group instead
-        for (var group : groups.entrySet()) {
-            var count = 0;
-            for (var groupEntry : group.getValue()) {
-                if (ret.contains(groupEntry)) {
-                    count++;
-                    if (count == 2) {
-                        ret.add(group.getKey());
-                        ret.removeAll(group.getValue());
-                        break;
-                    }
-                }
-            }
-        }
-        return ret;
-    }
-
-    public Set<String> allowed() {
-        return allowed;
-    }
-
-    public boolean isAllowed(String s) {
-        return allowed.contains(s);
-    }
+public interface LabelConfiguration {
+    Set<String> label(Set<Path> changes);
+    Set<String> allowed();
+    boolean isAllowed(String s);
 }
